Move domain-logic for createOrUpdate from kafka-consumer to service
