+++ src/main/kotlin/no/nav/syfo/personoppgavehendelse/kafka/KPersonoppgavehendelse.kt
@@ -4,10 +4,5 @@ import no.nav.syfo.personstatus.domain.OversikthendelseType
 
 data class KPersonoppgavehendelse(
     val personident: String,
-    val hendelsetype: String,
+    val hendelsetype: OversikthendelseType,
 )
-
-fun KPersonoppgavehendelse.oversikthendelseType(): OversikthendelseType? =
-    OversikthendelseType.values().firstOrNull {
-        it.name == this.hendelsetype
-    }
