+++ src/main/kotlin/no/nav/syfo/personoppgavehendelse/kafka/KafkaPersonoppgavehendelseTask.kt
@@ -2,9 +2,9 @@ package no.nav.syfo.personoppgavehendelse.kafka
 
 import no.nav.syfo.application.kafka.KafkaEnvironment
 import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.application.database.database
 import no.nav.syfo.application.kafka.kafkaAivenConsumerConfig
 import no.nav.syfo.kafka.launchKafkaTask
+import no.nav.syfo.personstatus.PersonoversiktStatusService
 import no.nav.syfo.util.configuredJacksonMapper
 import org.apache.kafka.clients.consumer.ConsumerConfig
 import org.apache.kafka.common.serialization.Deserializer
@@ -15,8 +15,9 @@ val PERSONOPPGAVEHENDELSE_TOPIC = "teamsykefravr.personoppgavehendelse"
 fun launchKafkaTaskPersonoppgavehendelse(
     applicationState: ApplicationState,
     kafkaEnvironment: KafkaEnvironment,
+    personoversiktStatusService: PersonoversiktStatusService,
 ) {
-    val personoppgavehendelseService = PersonoppgavehendelseService(database = database)
+    val personoppgavehendelseConsumerService = PersonoppgavehendelseConsumerService(personoversiktStatusService)
     val consumerProperties = Properties().apply {
         putAll(kafkaAivenConsumerConfig(kafkaEnvironment = kafkaEnvironment))
         this[ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG] =
@@ -26,7 +27,7 @@ fun launchKafkaTaskPersonoppgavehendelse(
         applicationState = applicationState,
         topic = PERSONOPPGAVEHENDELSE_TOPIC,
         consumerProperties = consumerProperties,
-        kafkaConsumerService = personoppgavehendelseService,
+        kafkaConsumerService = personoppgavehendelseConsumerService,
     )
 }
