+++ src/test/kotlin/no/nav/syfo/personstatus/KafkaOppfolgingstilfellePersonServiceSpek.kt
@@ -8,7 +8,6 @@ import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.domain.Virksomhetsnummer
 import no.nav.syfo.oppfolgingstilfelle.kafka.KafkaOppfolgingstilfelle
 import no.nav.syfo.personoppgavehendelse.kafka.KPersonoppgavehendelse
-import no.nav.syfo.personoppgavehendelse.kafka.PersonoppgavehendelseService
 import no.nav.syfo.personstatus.db.*
 import no.nav.syfo.personstatus.domain.OversikthendelseType
 import no.nav.syfo.testutil.*
@@ -24,7 +23,6 @@ import org.apache.kafka.clients.consumer.ConsumerRecords
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
 import java.time.Duration
-import java.util.UUID
 
 @InternalAPI
 object KafkaOppfolgingstilfellePersonServiceSpek : Spek({
@@ -33,6 +31,7 @@ object KafkaOppfolgingstilfellePersonServiceSpek : Spek({
         start()
 
         val externalMockEnvironment = ExternalMockEnvironment.instance
+        val internalMockEnvironment = InternalMockEnvironment.instance
         val database = externalMockEnvironment.database
 
         application.testApiModule(
@@ -41,7 +40,7 @@ object KafkaOppfolgingstilfellePersonServiceSpek : Spek({
 
         val kafkaOppfolgingstilfellePersonService = TestKafkaModule.kafkaOppfolgingstilfellePersonService
 
-        val personoppgavehendelseService = PersonoppgavehendelseService(database)
+        val personoversiktStatusService = internalMockEnvironment.personoversiktStatusService
 
         val mockKafkaConsumerOppfolgingstilfellePerson = TestKafkaModule.kafkaConsumerOppfolgingstilfellePerson
 
@@ -161,16 +160,11 @@ object KafkaOppfolgingstilfellePersonServiceSpek : Spek({
 
                 val oversiktHendelseOPLPSBistandMottatt = KPersonoppgavehendelse(
                     personIdentDefault.value,
-                    OversikthendelseType.OPPFOLGINGSPLANLPS_BISTAND_MOTTATT.name,
+                    OversikthendelseType.OPPFOLGINGSPLANLPS_BISTAND_MOTTATT,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversiktHendelseOPLPSBistandMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(oversiktHendelseOPLPSBistandMottatt)
+                )
 
                 kafkaOppfolgingstilfellePersonService.pollAndProcessRecords(
                     kafkaConsumer = mockKafkaConsumerOppfolgingstilfellePerson,
