+++ src/test/kotlin/no/nav/syfo/personstatus/api/v2/BehandlerdialogPersonoversiktStatusApiV2Spek.kt
@@ -7,7 +7,6 @@ import io.ktor.server.testing.*
 import io.ktor.util.*
 import no.nav.syfo.domain.PersonIdent
 import no.nav.syfo.personoppgavehendelse.kafka.KPersonoppgavehendelse
-import no.nav.syfo.personoppgavehendelse.kafka.PersonoppgavehendelseService
 import no.nav.syfo.personstatus.domain.OversikthendelseType
 import no.nav.syfo.testutil.*
 import no.nav.syfo.testutil.UserConstants.NAV_ENHET
@@ -17,7 +16,6 @@ import no.nav.syfo.util.*
 import org.amshove.kluent.shouldBeEqualTo
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
-import java.util.*
 
 @InternalAPI
 object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
@@ -28,8 +26,9 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
         with(TestApplicationEngine()) {
             start()
             val externalMockEnvironment = setupExternalMockEnvironment(application)
+            val internalMockEnvironment = InternalMockEnvironment.instance
             val database = externalMockEnvironment.database
-            val personoppgavehendelseService = PersonoppgavehendelseService(database)
+            val personoversiktStatusService = internalMockEnvironment.personoversiktStatusService
             val validToken = generateJWT(
                 audience = externalMockEnvironment.environment.azure.appClientId,
                 issuer = externalMockEnvironment.wellKnownVeilederV2.issuer,
@@ -44,16 +43,12 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
             it("return person with behandlerdialog_ubehandlet true when svar mottatt") {
                 val oversikthendelseBehandlerdialogSvarMottatt = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_SVAR_MOTTATT.name,
+                    OversikthendelseType.BEHANDLERDIALOG_SVAR_MOTTATT,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogSvarMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(oversikthendelseBehandlerdialogSvarMottatt)
+                )
+
                 database.setTildeltEnhet(
                     ident = PersonIdent(UserConstants.ARBEIDSTAKER_FNR),
                     enhet = NAV_ENHET,
@@ -77,16 +72,11 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
             it("return person with behandlerdialog_ubehandlet true when ubesvart melding mottatt") {
                 val oversikthendelseBehandlerdialogUbesvartMottatt = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_MELDING_UBESVART_MOTTATT.name,
+                    OversikthendelseType.BEHANDLERDIALOG_MELDING_UBESVART_MOTTATT,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogUbesvartMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(oversikthendelseBehandlerdialogUbesvartMottatt)
+                )
                 database.setTildeltEnhet(
                     ident = PersonIdent(UserConstants.ARBEIDSTAKER_FNR),
                     enhet = NAV_ENHET,
@@ -110,25 +100,18 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
             it("return person with behandlerdialog_ubehandlet true when ubesvart behandlet and svar mottatt at the same time") {
                 val oversikthendelseBehandlerdialogUbesvartBehandlet = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_MELDING_UBESVART_BEHANDLET.name,
+                    OversikthendelseType.BEHANDLERDIALOG_MELDING_UBESVART_BEHANDLET,
                 )
                 val oversikthendelseBehandlerdialogSvarMottatt = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_SVAR_MOTTATT.name,
+                    OversikthendelseType.BEHANDLERDIALOG_SVAR_MOTTATT,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogUbesvartBehandlet,
-                        callId = UUID.randomUUID().toString(),
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(
+                        oversikthendelseBehandlerdialogUbesvartBehandlet,
+                        oversikthendelseBehandlerdialogSvarMottatt,
                     )
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogSvarMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                )
                 database.setTildeltEnhet(
                     ident = PersonIdent(UserConstants.ARBEIDSTAKER_FNR),
                     enhet = NAV_ENHET,
@@ -152,16 +135,11 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
             it("return person with behandlerdialog_ubehandlet true when avvist melding mottatt") {
                 val oversikthendelseBehandlerdialogAvvistMottatt = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_MELDING_AVVIST_MOTTATT.name,
+                    OversikthendelseType.BEHANDLERDIALOG_MELDING_AVVIST_MOTTATT,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogAvvistMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(oversikthendelseBehandlerdialogAvvistMottatt)
+                )
                 database.setTildeltEnhet(
                     ident = PersonIdent(UserConstants.ARBEIDSTAKER_FNR),
                     enhet = NAV_ENHET,
@@ -185,16 +163,11 @@ object BehandlerdialogPersonoversiktStatusApiV2Spek : Spek({
             it("return no person when avvist melding behandlet") {
                 val oversikthendelseBehandlerdialogAvvistMottatt = KPersonoppgavehendelse(
                     UserConstants.ARBEIDSTAKER_FNR,
-                    OversikthendelseType.BEHANDLERDIALOG_MELDING_AVVIST_BEHANDLET.name,
+                    OversikthendelseType.BEHANDLERDIALOG_MELDING_AVVIST_BEHANDLET,
                 )
-                database.connection.use {
-                    personoppgavehendelseService.processPersonoppgavehendelse(
-                        connection = it,
-                        kPersonoppgavehendelse = oversikthendelseBehandlerdialogAvvistMottatt,
-                        callId = UUID.randomUUID().toString(),
-                    )
-                    it.commit()
-                }
+                personoversiktStatusService.createOrUpdatePersonoversiktStatuser(
+                    personoppgavehendelser = listOf(oversikthendelseBehandlerdialogAvvistMottatt)
+                )
                 database.setTildeltEnhet(
                     ident = PersonIdent(UserConstants.ARBEIDSTAKER_FNR),
                     enhet = NAV_ENHET,
